Obstacle Added
-Class obstacle added

+++ FinalGraficos1/FinalGraficos1/DeltaTime.cpp
@@ -18,7 +18,7 @@ void DeltaTime::Update(){
 	deltaclock.restart();
 }
 
-double DeltaTime::Get(){
+float DeltaTime::Get(){
 	return deltaTime;
 }
